==> 04_data-preprocessing.ipynb <==
# Final Project: Fake News Detection

By Felix Daubner - Hochschule der Medien

Module 'Supervised and Unsupervised Learning' - Prof. Dr.-Ing. Johannes Maucher

## Data Preprocessing

import pandas as pd
import numpy as np

from langdetect import detect

import nltk
nltk.download("stopwords")
nltk.download("punkt")
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize

from imblearn.over_sampling import RandomOverSampler
from gensim.models import KeyedVectors

import pickle

def replaceChannel(data, channel, alternative=None):
    '''
    This function replaces a string which contains varable 'channel' to just 'channel'. 
    An 'alternative' can also be given to search the string for 'alternative' and then replace it with given 'channel'.
    Returns Series.
    '''
    data["channel"] = data["channel"].replace(fr'.*{channel}.*', channel, regex=True)
    
    if alternative:
        data["channel"] = data["channel"].replace(fr'.*{alternative}.*', channel, regex=True)

    return data["channel"]

After getting some insights into the data, the data needs to be prepared to being able to train a machine learning model. Based on [data understanding](03_data-understanding.ipynb) some preprocessing were already discovered and will be taken on in the scope of this notebook.
Additionally, some other preprocessing steps have to be implemtend, such as:

- Remove 'full-flip', 'half-flip', 'no-flip' from target variable
- Control language of all statements and only keep English statements
- Handle missing values in column 'channel'
- Summarize and encode 'channel', for e.g. 'in a X post' and 'in a post on X' should become 'X'
- Drop column 'person' and decide about 'issue'
- Binarize and encode truth-column
- Balance dataset
- Convert text (string) to tokens
- Pad all sequences to the same length

All those operations have to be done on the test data (LIAR.csv) too, only using the insights we get from the training data.

First, import the raw data by using 'read_csv'.

data = pd.read_csv("data/scraped.csv", sep=";", index_col=0)

data.head()

### Remove invalid truth classifications

As already seen in [data understanding](03_data-understanding.ipynb) the target variable 'truth' not only contains the known and useable truth classifications which shall be predicted but also contains some variables which indicate whether a person has switched its views on a statement based on previous statements. Those switching indications can not be used and shall therefore be dropped from the dataset.

data.shape

data["truth"].unique()

The 'flip' list contains all variables which shall be removed as they do not indicate a classification of truth.

flip = ["half-flip", "full-flop", "no-flip"]

All rows containing in instance of 'flip' in column 'truth' are removed.

data_removedflip = data[~data["truth"].isin(flip)]

By comparing the resulting shape to the original shape, onyl 160 rows were removed from the dataset. This means that the majority of all rows  still remain part of the dataset and can be used for training a fake-news model.

data_removedflip.shape

print(f"Removed columns: {data.shape[0]} - {data_removedflip.shape[0]} = {data.shape[0] - data_removedflip.shape[0]}")

### Check language of statements

As seen in [data understanding](03_data-understanding.ipynb) there might be some rows which contain statements non-English statements. The fake-news model to be trained in the scope of this project is trained only on English statements. 

Therefore, it has to be checked whether all statements are in English. All non-English statements shall be dropped from the DataFrame. To check the language of a statement, the library 'langdetect' is used. 'langdetect' provides a method 'detect' which returns the language of a string. 'detect' is applied on the column 'statement' of the data, so every row is checked for the language of the statement. The result is saved in column 'lang'.

data_lang = data_removedflip.copy()

data_lang["lang"] = data_removedflip["statement"].apply(detect)

The following output shows the count of rows for every detected language. Based on 'langdetect' from all remaining 16766 rows only 16176 were in English. Nearly 600 rows are detected to be from other languages, the most common being Spanish.

data_lang["lang"].value_counts()

All rows which were detected to be non-English should be removed from the DataFrame without further checking. As the majority (about 94%) of all statements are classified to be English, not much data is lost. Still, more rows are dropped than expected.

data_eng = data_lang[data_lang["lang"] == "en"]

data_eng.shape[0] == sum(data_lang["lang"] == "en")

As column 'lang' is not further needed for processing, it can be dropped from the data.

data_cleanedlang = data_eng.drop(["lang"], axis=1)

### Handle missing values

The info()-method outputs some basic info of the DataFrame. As seen in the output below, column 'channel' still misses some values. There are many ways how to deal with missing values.

data_cleanedlang.info()

Two options are considered: either deleting the rows and therefore reduce the dataset or fill the missing values with 'Other' and keep the rows. Both options do not contribute in gaining information, but at least the second option does not lose information. So this procedure is preferred.

data_mv = data_cleanedlang.fillna("Other")

After filling the empty rows of column 'channel' all data is now complete.

data_mv.info()

### Summarize 'channel'

As already seen in [data understanding](03_data-understanding.ipynb), there are different channels the statements were made on. Unfortunately, those channel are not standardized in any way.

len(data_mv["channel"].unique())

The previous output displays the number of different channels. This number is quite high and should be reduced to a reasonable amount of channels. Based on previous insights in [data understanding](03_data-understanding.ipynb) the most important channels can already be defined.

channels = ["Facebook", "Instagram", "TikTok", "Threads", "X", "interview", "speech", "ad", "debate", "blog", "press", "campaign", 
            "TV", "radio", "video", "social media", "article", "lecture", "talk", "presentation", "mail", "podcast"]

A function 'replaceChannel' was implemented to scan through the given column and replace the content of every cell containing a certain keyword with the keyword for this respective channel. So for e.g. every cell containing the word 'Facebook' in all possible variations should be replaced by solely the word 'Facebook'. This is done for all channels which are already defined. All other cells are replaced by "Other" as the channel might be too specific.

data_channel = data_mv.copy()

for ch in channels:
    data_channel["channel"] = replaceChannel(data_channel, ch)

    if ch == "X":
        data_channel["channel"] = replaceChannel(data_channel, ch, "Twitter")
        data_channel["channel"] = replaceChannel(data_channel, ch, "twitter")
        data_channel["channel"] = replaceChannel(data_channel, ch, "Tweet")
        data_channel["channel"] = replaceChannel(data_channel, ch, "tweet")
        data_channel["channel"] = replaceChannel(data_channel, ch, "x")
    if ch == "social media":
        data_channel["channel"] = replaceChannel(data_channel, ch, "post")
    if ch == "TV":
        data_channel["channel"] = replaceChannel(data_channel, ch, "CNN")
        data_channel["channel"] = replaceChannel(data_channel, ch, "Fox")
    if ch == "campaign":
        data_channel["channel"] = replaceChannel(data_channel, ch, "rally")
        data_channel["channel"] = replaceChannel(data_channel, ch, "Campaign")
    if ch == "speech":
        data_channel["channel"] = replaceChannel(data_channel, ch, "Speech")
    if ch == "press":
        data_channel["channel"] = replaceChannel(data_channel, ch, "newsletter")

data_channel.loc[~data_channel["channel"].isin(channels), "channel"] = "Other"


Now, only the channels that have been defined earlier are left in column 'channel'.

data_channel["channel"].unique().tolist()

The count of all channels shows the most commonly used channels of statements. Especially social media like Facebook, X and Instagram but also interviews, ads and speeches appear very often. 

data_channel[["channel"]].groupby("channel").size()

Currently, the column 'channel' contains all the channels listed in the output above. The channels are represented by strings which are not usable in a machine learning model so the string have to be converted into a numerical representations. This can be done by one-hot-encoding which creates a column for every characteristic of 'channel' and assigns the number '1' only to the column of the channel the statement was made on. All other channels receive the value '0'. 

data_channel_enc = pd.get_dummies(data_channel, columns=["channel"], prefix="channel", drop_first=True)

As a result of one-hot-encoding there are many more columns compared to the original DataFrame structure. Also, the first column of the newly created encoded columns is deleted.

print(f"Columns (original structure): {data_channel.shape[1]}")
print(f"Columns (encoded structure): {data_channel_enc.shape[1]}")

data_channel_enc.columns.to_list()

### Handle 'person' and 'issue'

In this section it is evaluated whether the column 'person' can be preprocessed in any way.

The dataset contains 2765 different persons which a bit less than the number of channels before preprocessing but still too many to extract useful information.

len(data_channel_enc["person"].unique())

One way of 'cleaning' this column can be to replace the name of person with the party they are associated with. Even though the sample size was quite small in the previous notebook [data understanding](03_data-understanding.ipynb), there was a connection between member of a party and the probability of true/false statements. Member of the Republican party tend to have a higher probability of saying false statements, while members of the Democratic party tend to have a higher chance of saying true statements.
So coming back to preprocessing, this would mean that for e.g. the name 'Donald Trump' would be replaced with 'Republican', the name 'Barack Obama' would be replaced with 'Democrat'. 

Still, 2765 persons are too much to manually assign them to a party. Reliable sources are needed to assign a party to every person in the dataset. As the names itself do not contribute in assigning a person to a party, classical preprocessing steps can not be used.
Therefore, 'person' should not be considered as a feature and is dropped from the dataset.

However, 'issue' does contain only 150 characteristics which can be handled more easily than 2764 characteristics of 'person'. Still it is evaluated how much 'issue' can contribute in predicting fake-news. Currently, it is not clear whether 'issue' contains all possible issues which occur in the scope of this project or above. As seen in [data understanding](03_data-understanding.ipynb), some issues were only temporary like hurricane katrina or 2024 elections. Issues might be a useful indicator using a larger dataset and more generalized issues. Therefore, also 'issue' is dropped from the DataFrame.

data_dropped = data_channel_enc.drop(["person", "issue"], axis=1)

### Binarize 'truth'

Now, the target variable 'truth' should be converted to boolean value. As already seen in [data understanding](03_data-understanding.ipynb), the target variable 'truth' can take on different characteristics. Specifically, it can take on one of the values of 'pants-fire', 'false', 'barely-true', 'half-true', 'mostly-true' and 'true'. The machine learning model to be trained should only predict whether a statement is rather true or rather false, so the target variable has to be binarized. 

The following allocation will be considered:

true = ["true", "mostly-true", "half-true"]
false = ["barely-true", "false", "pants-fire"]

The target variable 'truth' is now replaced based on the defined lists for true and false values above. Statements which are classified in the 'true' range, receive value 1. Statements which are classified in the 'false' range, receive value 0.

data_binary = data_dropped.copy()
data_binary["truth"] = data_binary["truth"].replace(true, 1)
data_binary["truth"] = data_binary["truth"].replace(false, 0)

As seen below, the values got replaced and binarized in the same step. 

Also, both classes are not distributed evenly. This issue has to be taken on in the next step as it is required to balance out the dataset to prevent a majority class bias in the model.

data_binary["truth"].value_counts()

### Balance dataset

There are two popular ways to balance out a dataset to achieve an even distribution between two classes.

First option is to oversample. Oversampling means to copy random entries from the class with less occurences (minority class) until there are as many samples in both classes. The advantage when using oversampling is that no data is lost. Unfortunately, biases can be reproduced and even reinforced by multiplying samples.

Second options is to undersample. Undersampling is the opposite of oversampling. So, instead of multiplying samples of the minority class until both classes are even, this time samples from the class with more occurences (majority class) are deleted until both classes are balanced out. This leads to the loss of data which is not a good option to consider as we might lose out on relevant information for the fake-news model. 

So oversampling is used to balance out the data.

ros = RandomOverSampler(random_state=42)

Before starting to oversample, the data has to be splitted into X and y. X containing all features, y containing only the target variable 'truth'.

X = data_binary.drop(["truth"], axis=1)
y = data_binary["truth"]

Currently, there are 6095 samples of class '1' (or in this context: true statements), while there are 10085 samples of class '0' (false statements)

y.value_counts()

Now, the data is oversampled randomly with replacement. That means that from the minority class '1', a random sample is chosen and copied until. Then another sample is chosen randomly. This proceeds until both classes are balanced out.

X_resampled, y_resampled = ros.fit_resample(X, y)

After oversampling, the dataset is now balanced out and can be further prepared for training.

y_resampled.value_counts()

data_balanced = X_resampled.copy()

data_balanced["truth"] = y_resampled

### Tokenization of 'statement'

To make the statements usable for a machine learning model, the strings have to be converted into a numerical representation. Tokenization assigns every word a unique integer value. The tokenizer learns the mapping of words to number based on given texts and then applies the mapping on all words.

For tokenization to work, all statements have to be prepared. The strings need to be in lower case without special characters.

data_balanced["statement"] = data_balanced["statement"].str.lower().str.replace(r'[^a-zA-Z0-9 ]',"", regex=True).astype("str")

data_balanced[["statement", "truth"]].head()

from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.preprocessing.sequence import pad_sequences

Now, it is assumed that there are only English statements left in the data. Thus, the statements can now be tokenized using keras.Tokenizer which transforms the statements into a list of integers.

Two approaches are implemented. The first approach does not remove stop words. This results is saved in column 'token'. Also, a second approach where stop words are removed is executed and saved in column 'token_stop'. In the following process of this project, it is to be decided whether including or excluding stop words contributes to a better model performance.

NUM_WORDS limits the vocabulary of the tokenizer. For this project, NUM_WORDS is set to 3000 meaning only the 3000 most common words are kept in the tokenizers vocabulary and then transformed into an integer value. All other words are discarded and will not be used further. The tokenizer learns the words based on the statements contained in the data.

NUM_WORDS=3000

data_token = data_balanced.copy()

token = Tokenizer(num_words=NUM_WORDS)
statements = data_balanced["statement"].to_list()
token.fit_on_texts(statements)
data_token["token"] = token.texts_to_sequences(statements)

Same procedure is applied without the statements containing the stop words. The stop words are already defined in an imported object.

stop_words = set(stopwords.words("english"))

data_token["statement_stop"] = data_token["statement"].apply(lambda text: " ".join([word for word in text.split() if word not in stop_words]))

statements_stop = data_token["statement_stop"].to_list()
data_token["token_stop"] = token.texts_to_sequences(statements_stop)

The trained tokenizer is saved as a pickle object. Pickle objects can be imported into any other notebooks.

with open("tokenizer/tokenizer.pickle", "wb") as handle:
    pickle.dump(token, handle, protocol=pickle.HIGHEST_PROTOCOL)

By looking at the data, 'token' now contains the tokenized statements.

data_token[["statement", "token", "truth"]].head()

data_token[["statement_stop", "token_stop", "truth"]].head()

As seen in the previous output, not all list of tokens have the same size. To make all tokens the same size, the tokenized statements have to be padded with '0' until they all reach the same length.

All statements are padded to a length of the longest tokenized statement. The longest tokenized statement has a length of 57, so all tokenized statements are filled with 0 until a length of 57 is reached.

MAX_TOKEN = data_token["token"].apply(len).max()
MAX_TOKEN

### Padding of sequences

All statemtents have to be padded to a uniformal length of 57.

padded = pad_sequences(data_token["token"].to_list())

data_padded = data_token.copy()
data_padded["token"] = padded.tolist()

Padding all tokenized statements can be done by applying 'pad_sequences' which fills the statements with zeros until a length of 57.

padded_stop = pad_sequences(data_token["token"].to_list())

data_padded["token_stop"] = padded_stop.tolist()

After padding the tokenized statements, there are many zeros at the beginning. This is called pre-padding.

data_padded[["statement", "token", "truth"]].head()

data_padded[["statement_stop", "token_stop", "truth"]].head()

### Save processed training data

After the data is now fully prepared for model training, it is saved as a json. The reason it is saved as a json rather than a csv is that a json when importing into a DataFrame still keeps the original data structure. 

data_padded.to_json("data/processed.json", orient="records", lines=True)

### Preprocessing of test data

The test data (LIAR.csv) is not processed in the same way the training data got processed as this data is handled like unseen data. Currently, the test data does not comply with the structure of the training data. However, to being able to use the test data, it has to be in a certain format and structure. In the following, the test data which will be taken from 'LIAR' dataset is prepared for usage in the fake-news model.

All the processing steps are done only based on the knowledge found out in the process of preparing the training data. This prevents implementing a bias which already includes knowledge from the test data which can and should not be known at this point.

test = pd.read_csv("data/LIAR.csv", sep=";", index_col=0, header=None)

test.head()

First, all columns which do not contain similar information as in the training set are dropped.

test_dropped = test.drop([1, 6, 7, 8, 9, 10, 11, 12, 13], axis=1)

To have a better overview over the data, the columns are renamed to comply with the training data structure.

test_dropped.columns = ["truth", "statement", "issue", "person", "channel"]

After renaming, the data looks familiar. The column names are similar, also the structure inside the columns seems like the structure of the original training data.

test_dropped.head()

Following steps have to be done to achieve the same structure and format as the training data:
- Drop non-English rows
- Drop 'person' and 'issue'
- Summarize and encode 'channel'
- Binarize 'truth'
- Tokenize and pad 'statement'

As already said, all those steps are done using the same methods as done with the training data.
To automate this process, a preprocessing pipeline should be implemented. Due to the time restriction and a different focus of this project, a preprocessing class is implemented which implements all steps from preprocessing the training data.

class preprocessData():
    def __init__(self, data):
        self.__data = data

    def orchestrate(self):
        print("Check Basics ...", sep=' ', end='', flush=True) 
        data_b = self.__checkBasics(self.__data)
        print(" Finished Basics!")
        print("Check Language ...", sep=' ', end='', flush=True) 
        data_l = self.__checkLanguage(data_b)
        print(" Finished Language!")
        print("Check 'person' ...", sep=' ', end='', flush=True) 
        data_p = self.__checkPerson(data_l)
        print(" Finished 'person'!")
        print("Check 'issue' ...", sep=' ', end='', flush=True) 
        data_i = self.__checkIssue(data_p)
        print(" Finished 'issue'!")
        print("Handle 'channel' ...", sep=' ', end='', flush=True) 
        data_c = self.__handleChannel(data_i)
        print(" Finished handling of 'channel'!")
        print("Binarize 'truth' ...", sep=' ', end='', flush=True) 
        data_t = self.__binarizeTruth(data_c)
        print(" Finished binarizing of 'truth'!")
        print("Tokenize 'statement' ...", sep=' ', end='', flush=True) 
        data_result = self.__tokenizeStatement(data_t)
        print(" Finished tokenizing of 'statement'!")

        print("Finished preprocessing of data!")
        return data_result

    def __checkBasics(self, data):
        data.dropna(axis=0, subset=["truth", "statement"])
        return data

    def __checkLanguage(self, data):
        data["lang"] = data["statement"].apply(detect)
        return data[data["lang"]=="en"].drop(["lang"], axis=1)

    def __checkPerson(self, data):
        try:
            return data.drop(["person"], axis=1)
        except:
            return data
        
    def __checkIssue(self, data):
        try:
            return data.drop(["issue"], axis=1)
        except:
            return data

    def __handleChannel(self, data):
        for ch in channels:
            data["channel"] = replaceChannel(data, ch)

            if ch == "X":
                data["channel"] = replaceChannel(data, ch, "Twitter")
                data["channel"] = replaceChannel(data, ch, "twitter")
                data["channel"] = replaceChannel(data, ch, "Tweet")
                data["channel"] = replaceChannel(data, ch, "tweet")
                data["channel"] = replaceChannel(data, ch, "x")
            if ch == "social media":
                data["channel"] = replaceChannel(data, ch, "post")
            if ch == "TV":
                data["channel"] = replaceChannel(data, ch, "CNN")
                data["channel"] = replaceChannel(data, ch, "Fox")
            if ch == "campaign":
                data["channel"] = replaceChannel(data, ch, "rally")
                data["channel"] = replaceChannel(data, ch, "Campaign")
            if ch == "speech":
                data["channel"] = replaceChannel(data, ch, "Speech")
            if ch == "press":
                data["channel"] = replaceChannel(data, ch, "newsletter")

        data.loc[~data["channel"].isin(channels), "channel"] = "Other"

        return pd.get_dummies(data, columns=["channel"], prefix="channel", drop_first=True)

    def __binarizeTruth(self, data):
        data["truth"] = data["truth"].replace(true, 1)
        data["truth"] = data["truth"].replace(false, 0)
        return data

    def __tokenizeStatement(self, data):
        statements = data["statement"].to_list()
        data["token"] = token.texts_to_sequences(statements)

        data["statement_stop"] = data["statement"].apply(lambda text: " ".join([word for word in text if word not in stop_words]))
        statements_stop = data["statement_stop"].to_list()
        data["token_stop"] = token.texts_to_sequences(statements_stop)

        padded = pad_sequences(data["token"].to_list(), maxlen=MAX_TOKEN)
        data["token"] = padded.tolist()

        padded_stop = pad_sequences(data["token_stop"].to_list(), maxlen=MAX_TOKEN)
        data["token_stop"] = padded_stop.tolist()

        return data

Now the test data can be preprocessed using 'preprocessData'-class and 'orchestrate'-method which executes all preprocessing steps in the correct order.

pipeline = preprocessData(test_dropped)
test_processed = pipeline.orchestrate()

The result is a fully usable test dataset which is also saved as a json.

test_processed.to_json("data/LIAR_processed.json", orient="records", lines=True)
