==> 05_baseline.ipynb <==
# Final Project: Fake News Detection

By Felix Daubner - Hochschule der Medien

Module 'Supervised and Unsupervised Learning' - Prof. Dr.-Ing. Johannes Maucher

## Baseline model - Logistic Regression

import pandas as pd
import numpy as np

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report

NUM_WORDS=3000
MAX_SEQUENCE_LEN = 57

To being able to compare the results of the neural networks to be trained, a simple baseline model will be implemented. The baseline models acts as a reference and will be implemented without further exploration, discussion and / or optimization.

As the task to solve is a classification task, a logistic regression is trained and evaluated on the already preprocessed data.

### Prepare data for model training

At first, the data which was preprocessed in the previous notebook is imported. Still, it needs to be adjusted to train a Logistic Regression model.

data = pd.read_json("data/processed.json", orient="records", lines=True)

The data is splitted into features (X) and target (y). As the model should be kept simple, only the tokenized text data is used. The encoded categorical columns are not used in this baseline model.

X = np.array(data["token"].apply(np.array).to_list())
y = np.array(data["truth"])

Now, the data is splitted into training and test data which prepares the data for the training and evaluation process.

X_train, X_test, y_train, y_test = train_test_split(X, y, train_size=0.7, random_state=42)

### Train model

Now, the model is initialized and then trained using only the training data. 
The features only consits of the tokenized words. Whether this is a useful model shall not be discussed. The performance expectations are at around 55 to 60% accuracy.

log = LogisticRegression(solver="newton-cholesky")
log.fit(X_train, y_train)

### Evaluate model

Some evaluations are done using first the training and then the test set. A classification report should provide some insights into the  performance of the model which will be the reference for the neural network.

First, the model shall predict the training data. The predictions are then saved in a column 'predicted' and compared with the true values of each instance.

results_train = pd.DataFrame(y_train, columns=["true"])
results_train["predicted"] = log.predict(X_train)

results_train["correct"] = results_train["true"] == results_train["predicted"]

As seen in the output below, there are more instances predicted as 'true' than as 'false'.

results_train[["correct"]].value_counts()

As expected, the accuracy of the model on the training data is at 57%. This means 57% of the training data was classified correctly after training the model with this data. 

The accuracy value is close to 50% which would be equal to just randomly predict a target value for each instance of the data.

print(f"Classification Report of training data:")
print(classification_report(results_train["true"], results_train["predicted"]))

Now, the performance on the test data is calculated using the Logistic regression model. The test data is unseen data which the model has not seen before.

results_test = pd.DataFrame(y_test, columns=["true"])
results_test["predicted"] = log.predict(X_test)

results_test["correct"] = results_test["true"] == results_test["predicted"]

Again, more instances have been classified as 'true' than 'false'.

results_test[["correct"]].value_counts()

When looking at the results of the test data, the accuracy is slightly but not significantly worse compared to the training data. An accuracy of 56% is slightly better than a random prediction but definitely not good enough.

print(f"Classification Report of test data:")
print(classification_report(results_test["true"], results_test["predicted"]))
