==> 07_sentiment.ipynb <==
# Final Project: Fake News Detection

By Felix Daubner - Hochschule der Medien

Module 'Supervised and Unsupervised Learning' - Prof. Dr.-Ing. Johannes Maucher

## Sentiment Analysis

import pandas as pd
import numpy as np
import altair as alt

from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report

from tensorflow import keras
from keras.models import Model
from keras.layers import Embedding, Flatten, Dense, LSTM, Conv1D, Flatten, MaxPooling1D, Dropout, Bidirectional, Input, Concatenate

import pickle
from gensim.models import KeyedVectors

from transformers import AutoModelForSequenceClassification, AutoTokenizer, Trainer

NUM_WORDS=3000
MAX_SEQUENCE_LEN = 57
NUM_CAT = 20
MODEL = "siebert/sentiment-roberta-large-english"

sentiment = AutoModelForSequenceClassification.from_pretrained(MODEL)

tokenizer = AutoTokenizer.from_pretrained(MODEL)
trainer = Trainer(model=sentiment)
